requirements.txt
torch>=1.12.0
torchaudio>=0.12.0
numpy>=1.21.0
librosa>=0.9.0
soundfile>=0.10.0
scipy>=1.7.0
matplotlib>=3.5.0
tqdm>=4.64.0
PyYAML>=6.0
click>=8.0.0
colorama>=0.4.0
PyQt5>=5.15.0
pyqtgraph>=0.12.0
modelscope>=1.9.0
transformers>=4.20.0 